refactoring table
