+++ package.json
@@ -8,7 +8,9 @@
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^13.5.0",
         "bootstrap": "^5.1.0",
+        "bootstrap-icons": "^1.11.1",
         "lodash": "^4.17.15",
+        "prettier": "^3.1.0",
         "prop-types": "^15.8.1",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
