+++ src/app/components/products.jsx
@@ -10,13 +10,23 @@ import ProductsTable from "./productsTable";
 
 import _ from "lodash";
 
-const Products = ({ products: allProducts, ...rest }) => {
+const Products = () => {
     const [currentPage, setCurrentPage] = useState(1);
     const [categories, setCategories] = useState();
     const [selectedCategory, setSelectedCategory] = useState();
     const [sortBy, setSortBy] = useState({ iter: "price", order: "asc" });
-
     const pageSize = 4;
+
+    const [products, setProducts] = useState();
+    console.log("products App.jsx", products);
+
+    useEffect(() => {
+        api.products.fetchAll().then((data) => setProducts(data));
+    }, []);
+
+    const handleClick = (prodId) => {
+        setProducts(products.filter((product) => product.id === prodId));
+    };
 
     useEffect(() => {
         api.categories.fetchAll().then((data) => setCategories(data));
@@ -42,74 +52,77 @@ const Products = ({ products: allProducts, ...rest }) => {
         console.log(item);
     };
 
-    const filteredProducts = selectedCategory
-        ? allProducts.filter(
-              (product) =>
-                  JSON.stringify(product.category) ===
-                  JSON.stringify(selectedCategory)
-          )
-        : allProducts;
+    if (products) {
+        const filteredProducts = selectedCategory
+            ? products.filter(
+                  (product) =>
+                      JSON.stringify(product.category) ===
+                      JSON.stringify(selectedCategory)
+              )
+            : products;
 
-    const count = filteredProducts.length;
+        const count = filteredProducts.length;
 
-    const sortedProducts = _.orderBy(
-        filteredProducts,
-        [sortBy.path],
-        [sortBy.order]
-    );
+        const sortedProducts = _.orderBy(
+            filteredProducts,
+            [sortBy.path],
+            [sortBy.order]
+        );
 
-    const productCrop = paginate(sortedProducts, currentPage, pageSize);
-    console.log(productCrop);
+        const productCrop = paginate(sortedProducts, currentPage, pageSize);
+        console.log(productCrop);
 
-    const clearFilter = () => {
-        setSelectedCategory();
-    };
+        const clearFilter = () => {
+            setSelectedCategory();
+        };
 
-    return (
-        <>
-            <SearchStatus length={count} />
-            <SearchStatus length={allProducts.length} />
-            <div className="d-flex">
-                {categories && (
-                    <div className="d-flex flex-column flex-shrink-0 p-3">
-                        <GroupList
-                            selectedItem={selectedCategory}
-                            items={categories}
-                            onItemSelect={handleCategoriesSelect}
+        return (
+            <>
+                <SearchStatus length={count} />
+                <SearchStatus length={products.length} />
+                <div className="d-flex">
+                    {categories && (
+                        <div className="d-flex flex-column flex-shrink-0 p-3">
+                            <GroupList
+                                selectedItem={selectedCategory}
+                                items={categories}
+                                onItemSelect={handleCategoriesSelect}
+                            />
+                            <button
+                                className="btn btn-secondary mt-2"
+                                onClick={clearFilter}
+                            >
+                                Очистить
+                            </button>
+                        </div>
+                    )}
+                    {/* <div className="container"> */}
+                    <div className="d-flex flex-column flex-shrink-0">
+                        <ProductsTable
+                            products={productCrop}
+                            onSort={handleSort}
+                            selectedSort={sortBy}
+                            handleClick={handleClick}
                         />
-                        <button
-                            className="btn btn-secondary mt-2"
-                            onClick={clearFilter}
-                        >
-                            Очистить
-                        </button>
-                    </div>
-                )}
-                {/* <div className="container"> */}
-                <div className="d-flex flex-column flex-shrink-0">
-                    <ProductsTable
-                        products={productCrop}
-                        onSort={handleSort}
-                        selectedSort={sortBy}
-                        {...rest}
-                    />
-                    <div className="d-flex justify-content-center">
-                        <Pagination
-                            itemsCount={count}
-                            pageSize={pageSize}
-                            currentPage={currentPage}
-                            onPageChange={handlePageChange}
-                        />
+                        <div className="d-flex justify-content-center">
+                            <Pagination
+                                itemsCount={count}
+                                pageSize={pageSize}
+                                currentPage={currentPage}
+                                onPageChange={handlePageChange}
+                            />
+                        </div>
                     </div>
                 </div>
-            </div>
-            {/* </div> */}
-        </>
-    );
+                {/* </div> */}
+            </>
+        );
+    }
+    return "Loading products.jsx";
 };
 
 Products.propTypes = {
-    products: PropTypes.array.isRequired
+    products: PropTypes.array
 };
 
 export default Products;
